requirements.txt
# Telegram Bot Framework
python-telegram-bot==13.15

# Firebase (replacing Supabase)
firebase-admin==6.4.0

# Environment Variables
python-dotenv==1.0.0

# Google Gemini API
google-generativeai==0.3.1

# PDF Generation
WeasyPrint==60.2
Jinja2==3.1.2
markupsafe==2.1.3

# HTTP Requests
requests==2.31.0

# Utilities
python-dateutil==2.8.2